leetcode/0095.UniqueBinarySearchTreesII/95. Unique Binary Search Trees II.go
package leetcode

type TreeNode struct {
	Val   int
	Left  *TreeNode
	Right *TreeNode
}

func generateTrees(n int) []*TreeNode {
	if n == 0 {
		return nil
	}
	return generateTreesFunc(1, n)
}

func generateTreesFunc(start, end int) []*TreeNode {
	tree := []*TreeNode{}
	if start > end {
		tree = append(tree, nil)
		return tree
	}
	for i := start; i <= end; i++ {
		left := generateTreesFunc(start, i-1)
		right := generateTreesFunc(i+1, end)
		for _, l := range left {
			for _, r := range right {
				root := &TreeNode{Val: i, Left: l, Right: r}
				tree = append(tree, root)
			}

		}
	}
	return tree
}
